# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateSemesterRequest:
    properties:
      term:
        enum:
        - spring
        - fall
        type: string
      year:
        maximum: 2200
        minimum: 1900
        type: integer
    required:
    - term
    - year
    type: object
  dto.CreateUniversityRequest:
    properties:
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
    required:
    - is_active
    - name_en
    - name_fa
    type: object
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          error message
          example: invalid request format
        type: string
    type: object
...
  dto.SemesterResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      term:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  dto.UniversityResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateSemesterRequest:
    properties:
      term:
        enum:
        - spring
        - fall
        type: string
      year:
        maximum: 2200
        minimum: 1900
        type: integer
    required:
    - term
    - year
    type: object
  dto.UpdateUniversityRequest:
    properties:
      is_active:
        type: boolean
      name_en:
        type: string
      name_fa:
        type: string
    required:
    - is_active
    - name_en
    - name_fa
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Termustat API
  version: "1.0"
paths:
...
  /v1/admin/semesters:
    get:
      description: Retrieves all semesters, ordered by most recent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SemesterResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Semesters
      tags:
      - semesters
    post:
      consumes:
      - application/json
      description: Creates a new semester with year and term
      parameters:
      - description: Create semester payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSemesterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SemesterResponse'
        "400":
          description: Invalid input or duplicate year+term
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Semester already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Semester
      tags:
      - semesters
  /v1/admin/semesters/{id}:
    delete:
      description: Deletes a semester by its ID
      parameters:
      - description: Semester ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Semester deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid semester ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Semester not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Semester
      tags:
      - semesters
    get:
      description: Retrieves a semester by its ID
      parameters:
      - description: Semester ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SemesterResponse'
        "400":
          description: Invalid semester ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Semester not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Semester
      tags:
      - semesters
    put:
      consumes:
      - application/json
      description: Updates the year or term of an existing semester
      parameters:
      - description: Semester ID
        in: path
        name: id
        required: true
        type: string
      - description: Update semester payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSemesterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SemesterResponse'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Semester not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Duplicate semester
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Semester
      tags:
      - semesters
  /v1/admin/universities:
    get:
      description: Retrieves all universities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UniversityResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Universities
      tags:
      - universities
    post:
      consumes:
      - application/json
      description: Creates a new university
      parameters:
      - description: Create university payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUniversityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UniversityResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict (name already exists)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create University
      tags:
      - universities
  /v1/admin/universities/{id}:
    delete:
      description: Deletes the specified university
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: University deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid university ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: University not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete University
      tags:
      - universities
    get:
      description: Retrieves a university by its ID
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UniversityResponse'
        "400":
          description: Invalid university ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: University not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get University
      tags:
      - universities
    put:
      consumes:
      - application/json
      description: Updates the specified university’s details
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: string
      - description: Update university payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUniversityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UniversityResponse'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: University not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update University
      tags:
      - universities
...
swagger: "2.0"